v1 Changelog: -Handled overflowing problem. -Specified conditions the gases are under. -Changed current to currNode. -Created deepCopy() method. -Added more test cases. -Handled item not in list case. -Handled printing empty list.

+++ src/personalworks/Node.java
@@ -3,11 +3,11 @@ package personalworks;
 public class Node {
     //Holds some form of data that is let's say number of collisions per second in a 1L of a real gas
     String gasName;
-    int collisions;
+    double collisions;
     Node next;
 
     //Constructor
-    public Node(String name, int collision){
+    public Node(String name, double collision){
         this.gasName = name;
         this.collisions = collision;
         this.next = null;
